Formatting improvements

+++ client/discord/events/messageDelete.ts
@@ -7,9 +7,10 @@ const event: Event = {
     async execute(message, { database, client }) {
         if (!message.partial && message.channel.type === "DM") return;
         const query = { message: message.id };
-        await database.discord.collection<Discord.panel>('Panels').deleteOne(query);
-        if (!message.partial && message.author.id !== client.user.id) return;
-        await database.discord.collection<Discord.giveaway>('Giveaways').deleteOne(query);
+        const panel = await database.discord.collection<Discord.panel>('Panels').deleteOne(query);
+        if (panel.deletedCount || (!message.partial && message.author.id !== client.user.id)) return;
+        const giveaway = await database.discord.collection<Discord.giveaway>('Giveaways').deleteOne(query);
+        if (giveaway.deletedCount) return;
         const notifications = database.discord.collection<Discord.notification>('Notifications');
         await notifications.updateOne({ webhooks: { $elemMatch: query } }, {
             $pull: {
